+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+### v1.0.1 (Build 20190609)
+* UPD: Plugin renamed to Indoor Air Quality Monitor (IAQM)
+* UPD: Changed IAQ Levels & Color according IAQ Standard
+* UPD: Changed Air Quality Device (Voltcraft) to Custom Device to match IAQ Index
+* UPD: IAQ Indicator (LED) brightness applies to all IAQ levels. Brightness 0 turns LED off
+* UPD: Python code in areas rewritten (use of arrays,  additional defs, simplified)
+* UPD: Various minor correction
+
 ### v1.0.0 (Build 20190608)
 * NEW: First version published